% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{attenuate}
\alias{attenuate}
\title{Helps with delaying and smoothing changes in R0}
\usage{
attenuate(x, steps = 5, type = "logistic")
}
\description{
The function smoothes numeric vectors either by logistic or linear
interpolation.
}
