% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_functions.R
\name{checkpoint.check.reload}
\alias{checkpoint.check.reload}
\title{Reload a checkpoint.}
\usage{
checkpoint.check.reload(ep, sp)
}
\arguments{
\item{ep}{An execution parameter list as decribed in \code{\link[=set.exec.params]{set.exec.params()}}.}

\item{sp}{A list with static model parameters as described in \code{\link[=set.static.params]{set.static.params()}}.}
}
\description{
The function loads data necessary to do a checkpoint restart and checks
them for usability and differences.
}
