% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_total.R
\docType{data}
\name{connect_total}
\alias{connect_total}
\title{Comuter matrix of complete population}
\format{
An object of class \code{"data.frame"} with  402 columns.
\enumerate{
\item \code{"dist_id"} - An integer representing the county's unique identifier.
\item \code{"Name"} - The name of the county.
\item \code{"Area"} - The counties area in km².
\item \code{"Inhabitants"} - The population of the county.
}
}
\source{
\href{https://statistik.arbeitsagentur.de/SiteGlobals/Forms/Suche/
Einzelheftsuche_Formular.html;jsessionid=98E7BF6228AC1E77E530233B76A8A14F?
nn=20934&topic_f=beschaeftigung-sozbe-krpend}{Direct link}
}
\usage{
data(connect_total)
}
\description{
....
}
\references{
German Federal Employment Agency, Pendlerverflechtungen der
sozialversicherungspflichtig Beschäftigten nach Ländern - Deutschland
(Jahreszahlen), (2019), \href{https://statistik.arbeitsagentur.de/DE/
Navigation/Statistiken/Themen-im-Fokus/Regionale-Mobilitaet/
Regionale-Mobilitaet-Nav.html}{https://statistik.arbeitsagentur.de}.
}
\keyword{datasets}
