% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_work.R
\docType{data}
\name{connect_work}
\alias{connect_work}
\title{Comuter matrix of working population}
\format{
An object of class \code{"data.frame"} with  402 columns.
\enumerate{
\item \code{"dist_id"} - An integer representing the county's unique identifier.
\item \code{"Name"} - The name of the county.
\item \code{"Area"} - The counties area in km².
\item \code{"Inhabitants"} - The population of the county.
}
}
\source{
...
}
\usage{
data(connect_work)
}
\description{
....
}
\references{
...
}
\keyword{datasets}
