% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_functions.R
\name{convert.Rp.to.Fp}
\alias{convert.Rp.to.Fp}
\title{Convert R-model parameters to Fortran-model input files}
\usage{
convert.Rp.to.Fp(filename, sp, iol, R0_effects, outpath = "./")
}
\arguments{
\item{filename}{Path to the output file.}

\item{sp}{A list with static model parameters as described in \code{\link[=set.static.params]{set.static.params()}}.}
}
\description{
The function prints the R-model parameter lists and input data to textfiles
which can be used as input for the Fortran model version.
}
