% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counties.R
\docType{data}
\name{counties}
\alias{counties}
\title{Structure of German counties}
\format{
An object of class \code{"data.frame"} with  four columns.
\enumerate{
\item \code{"dist_id"} - An integer representing the county's unique identifier.
\item \code{"Name"} - The name of the county.
\item \code{"Area"} - The counties area in km².
\item \code{"Inhabitants"} - The population of the county.
}
}
\source{
\href{https://www.destatis.de/DE/Themen/Laender-Regionen/
Regionales/Gemeindeverzeichnis/Administrativ/Archiv/Standardtabellen/
04_KreiseVorjahr.html}{Statistisches Bundesamt}
}
\usage{
data(counties)
}
\description{
The German county structure representing the NUTS-3 level for Germany and by
that the spatial simulation structure in the CoSMic default setup.
}
\references{
Federal Statistical Office of Germany. Kreisfreie Städte und
Landkreise nach Fläche, Bevölkerung und Bevölkerungsdichte, (2018),
\href{https://www.destatis.de/DE/Themen/Laender-Regionen/Regionales}
{https://www.destatis.de/DE/Themen/Laender-Regionen/Regionales}.
}
\keyword{datasets}
