% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.execution.R
\name{export.to.slaves}
\alias{export.to.slaves}
\title{Export Variables to slaves}
\usage{
export.to.slaves(ep, cl, varlist)
}
\arguments{
\item{ep}{Execution parameter list. Use \code{\link[=set.exec.params]{set.exec.params()}} in order to
create a valid layout.}

\item{cl}{A parallel cluster prepared by \code{\link[=init.parallel.execution]{init.parallel.execution()}}.}

\item{varlist}{Vector of character strings representing variable names to
be exported to \code{cl}'s workers.}
}
\description{
For convenience this function wraps the \code{exportDoMPI} and
\code{clusterExport} functions from the \code{doMPI} and \code{doParallel}
packages.
}
