% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoSMic.R
\name{ff}
\alias{ff}
\title{CoSMic model function wrapper}
\usage{
ff(x, ep, sp, iol, pspace, sim.struc, op)
}
\arguments{
\item{x}{Numeric vector with model parameters subject to optimization.}

\item{ep}{Execution parameter list. Use \code{\link[=set.exec.params]{set.exec.params()}} in order to
create a valid layout.}

\item{sp}{List with static model parameters. Use \code{\link[=set.static.params]{set.static.params()}} to
create a valid layout.}

\item{iol}{Input data list. Use \code{\link[=load.input]{load.input()}} to load needed fies and
\code{\link[=init.connectivity]{init.connectivity()}} in order to create a valid date layout.}

\item{pspace}{List holding the parameter space with potentially variable
model parameters. Use the setter function \code{\link[=set.pspace]{set.pspace()}} to
add parameters.}

\item{sim.struc}{List with population data. Use \code{\link[=init.spatial.population]{init.spatial.population()}}
in order to create a valid layout.}

\item{op}{List with steering parameters for the optimization process. \cr
Use \code{\link[=set.optimization.params]{set.optimization.params()}} in order to create a valid layout
and\cr \code{\link[=init.reference.data]{init.reference.data()}} in order to init the optimization
targets based on observed data.}
}
\value{
A scalar value calculated according to the settings given in op.
}
\description{
This function wraps the CoSMic model function so that it can be used in the
GA algorithm as the objective function.
}
