% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.execution.R
\name{finalize.parallel.execution}
\alias{finalize.parallel.execution}
\title{Finalize parallel execution execution.}
\usage{
finalize.parallel.execution(ep, cl)
}
\arguments{
\item{ep}{Execution parameter list. Use \code{\link[=set.exec.params]{set.exec.params()}} in order to
create a valid layout.}

\item{cl}{A parallel cluster prepared by \code{\link[=init.parallel.execution]{init.parallel.execution()}}.}
}
\description{
For convenience this function wraps the \code{closeCluster} and
\code{stopCluster} functions from the \code{doMPI} and \code{doParallel}
packages.
}
