% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{fres.to.dataframe}
\alias{fres.to.dataframe}
\title{Load Fortran results as data.frame}
\usage{
fres.to.dataframe(data.dir, basename)
}
\description{
The function converts the result files of the Fortran model version in
single model execution mode to data.frames.
}
