% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{ftrain.to.dataframe}
\alias{ftrain.to.dataframe}
\title{Load Fortran training results as data.frame}
\usage{
ftrain.to.dataframe(data.dir, basename, split.col = "SH1")
}
\description{
The function converts the result files of the Fortran model version in
training execution mode to data.frames.
}
