% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoSMic.R
\name{GA.Monitor}
\alias{GA.Monitor}
\title{GA algorithm monitoring function}
\usage{
GA.Monitor(
  obj,
  digits = getOption("digits"),
  sp.int = static.params,
  op.int = opt.params
)
}
\arguments{
\item{obj}{An object provided by the GA function.}

\item{digits}{The number of digits provided by \code{getOption("digits")}.}

\item{sp.int}{List with static model parameters as created by
\code{\link[=set.static.params]{set.static.params()}}.}

\item{op.int}{List with steering parameters for the optimization process
as created by \cr \code{\link[=set.static.params]{set.static.params()}}.}
}
\description{
The function provides intermediate output after each iteration of the GA
algorithm.
}
