% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.population.R
\name{init.connectivity}
\alias{init.connectivity}
\title{Initialize regional connectivity}
\usage{
init.connectivity(iol, sp, ss)
}
\arguments{
\item{iol}{Input data list. Use \code{\link[=load.input]{load.input()}} to load needed fies and
\code{\link[=init.connectivity]{init.connectivity()}} in order to create a valid date layout.}

\item{sp}{List with static model parameters. Use \code{\link[=set.static.params]{set.static.params()}} to
create a valid layout.}

\item{ss}{List with population data. Use \code{\link[=init.spatial.population]{init.spatial.population()}} in
order to create a valid layout.}
}
\value{
An input data list with modified \code{connect_work} and
\code{connect_total} components. See \code{\link[=load.input]{load.input()}} about details on
how the input data list has to be strucutred in order to be
correctly modified by this function.
}
\description{
The function initializes the regional connectivity matrix according to the
requested regions to simulate.
}
