% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoSMic.R
\name{init.lhc}
\alias{init.lhc}
\title{Prepare parameter space}
\usage{
init.lhc(pspace, sp)
}
\arguments{
\item{pspace}{The parameter list pspace set by repeated calls to \code{\link[=set.pspace]{set.pspace()}}}

\item{A}{list with static model parameters as described in \code{\link[=set.static.params]{set.static.params()}}.}
}
\value{
A data.frame with dimension
[<# different evaluations> x <potentially_variable_model_params>]
If all model parameters in pspace are fixed, i.e. not variable
dim(lhc) will be [sp$iter x <potentially_variable_model_params>]
}
\description{
The function initializes the data.frame carrying the different sets of model
parameters resulting from the parameter variations set in the pspace list.
}
