% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.execution.R
\name{init.parallel.execution}
\alias{init.parallel.execution}
\title{Initalization of the parallel execution.}
\usage{
init.parallel.execution(ep, sp = NULL, op = NULL)
}
\description{
The function prepares and initializes the parallel execution of the
\code{\link[=CoSMic]{CoSMic()}} model function on computer clusters in dependence from the
requested execution procedure and selected parallel execution method.
}
