% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.population.R
\name{init.reference.data}
\alias{init.reference.data}
\title{Initialization of reference data.}
\usage{
init.reference.data(iol, op, sp, sim.struc)
}
\arguments{
\item{iol}{Input data list. Use \code{\link[=load.input]{load.input()}} to load needed fies and
\code{\link[=init.connectivity]{init.connectivity()}} in order to create a valid date layout.}

\item{op}{List with steering parameters for the optimization process. \cr
Use \code{\link[=set.optimization.params]{set.optimization.params()}} in order to create a valid layout.}

\item{sp}{List with static model parameters. Use \code{\link[=set.static.params]{set.static.params()}} to
create a valid layout.}

\item{sim.struc}{List with population data. Use \code{\link[=init.spatial.population]{init.spatial.population()}}
in order to create a valid layout.}
}
\value{
The list with steering parameters for the optimization process passed
in as parameter \code{op} with an additional component
\code{opt.target} carying observed data, prepared to be used as
target data in the optimization procedure of the \code{\link[=CoSMic]{CoSMic()}}
function.
}
\description{
The function adds a component to the optimization parameter list passed in
as parameter \code{op}. The added component \code{opt.target} contains
observed data depending which data are provided on input to the function
\code{\link[=load.input]{load.input()}}. The function additionally checks whether execution of the
optimization procedure is possible based on the selected optimization
targets and the provided data.
}
\section{ToDo}{

Implement ot[\link{dea.nuts2}]
}

