% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.population.R
\name{init.spatial.population}
\alias{init.spatial.population}
\title{Initialization of the population and its spatial structure.}
\usage{
init.spatial.population(iol, sp)
}
\arguments{
\item{iol}{Input data list. Use \code{\link[=load.input]{load.input()}} to load needed fies and
\code{\link[=init.connectivity]{init.connectivity()}} in order to create a valid date layout.}

\item{sp}{List with static model parameters. Use \code{\link[=set.static.params]{set.static.params()}} to
create a valid layout.}
}
\value{
A list with population data. Structured as follows:\preformatted{ sim.struc : List of 3
           $ pop     : data.frame
                     $ dist_id: int
                     $ date   : chr  
                     $ sex    : chr
                     $ age_gr : int
                     $ total  : int
           $ counties: int 
           $ states  : data.frame
                     $ Code       : int
                     $ inhabitants: int
                     $ Shortcut   : chr
                     $ Name       : chr
}
}
\description{
The function initializes the population and its spatial structure according
to the layout requested by \code{sp$sim.regions}, i.e. the regions selected
either at county or state level to be simulated.
}
