% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_functions.R
\name{load.input}
\alias{load.input}
\title{Loading input data}
\usage{
load.input(
  data.dir,
  country,
  trans.pr,
  pop.data,
  inf.cases,
  dead.cases,
  connect.total,
  connect.work,
  states,
  counties,
  R0.matrix.inp = NULL,
  dead.cases.by.state,
  dead.cases.by.country,
  icu.cases.by.county,
  icu.cases.by.state,
  icu.cases.by.country,
  lhc.data = NULL
)
}
\description{
Loading input data
}
