% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_functions.R
\name{map.R0effects}
\alias{map.R0effects}
\title{Map R0effects from NUTS-1 to NUTS-2}
\usage{
map.R0effects(R0effect.nuts2, R0effect.states, rows = NULL)
}
\arguments{
\item{R0effect.nuts2}{R0efects to map to}

\item{R0effect.states}{R0efects to map from}

\item{rows}{How many rows to map}
}
\description{
The function maps R0effects on NUTS-1 i.e. German state level to R0effects
on NUTS-2 level.
}
