% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param.space.R
\docType{data}
\name{param.space}
\alias{param.space}
\title{Parameter space}
\format{
An data structure whose components can be initialized by calls to
\code{set.pspace}.
}
\usage{
data(param.space)
}
\description{
Data structure to store model parameters which are protentialy subject to
variations.
}
\keyword{datasets}
