% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-functions.R
\name{plot.R0effect}
\alias{plot.R0effect}
\title{Plot R0effects over R0changes}
\usage{
\method{plot}{R0effect}(R0effect, sp, outfile = NULL, silent = FALSE)
}
\description{
The function plots timelines of the R0effects per state or NUTS2 region.
}
