% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-functions.R
\name{plots.by.country}
\alias{plots.by.country}
\title{Plot timelines accross the complete country}
\usage{
plots.by.country(
  outfile,
  sp,
  seed_icu,
  seed_dea,
  iol,
  pspace,
  rr,
  ind.states = NULL,
  global.plot,
  x.min = NULL,
  x.max = NULL,
  relative = FALSE,
  silent = FALSE,
  split.in = NULL,
  y.max = NULL,
  prog = NULL
)
}
\description{
The function plots timelines accross the complete country. Either fully
aggregated with global.plot = TRUE or aggregated once across the first
column of the latin hypercube, across each direct parameter with more than
one value and once across the parameter set of the first directv parameter.
}
