% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-functions.R
\name{plots.by.state}
\alias{plots.by.state}
\title{Plot timelines accross each state}
\usage{
plots.by.state(
  outfile,
  sp,
  seed_icu,
  seed_dea,
  iol,
  pspace,
  rr,
  region,
  fix.lim,
  x.min = NULL,
  x.max = NULL,
  filtered = FALSE,
  fk.cases = rep(1/7, 7),
  Sec.Axis = "RMS",
  fk.sec = rep(1/15, 15),
  sec.text = FALSE,
  ind.states = NULL,
  silent = FALSE,
  relative = FALSE,
  split.in = NULL,
  y.max = NULL,
  prog = NULL,
  overlay = NULL
)
}
\description{
The function plots timelines accross each state, aggregated once across the
first column in the latin hypercube, once across each direct parameter with
more than one value and once across the parameter set of the first directv
parameter.
}
