% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-functions.R
\name{plots.fortran}
\alias{plots.fortran}
\title{################################################################################
Plot Results of CoSMic-Fortran}
\usage{
plots.fortran(sp, iol, pspace, input.dir)
}
\description{
The function is provided for convenience. It loads the results of CoSMic Fortran
version, converts them to R and plots them using \code{plots.by.country()} and
\code{plots.by.state()}.
}
