% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/181231_pop_age_sex_distr_ger.R
\docType{data}
\name{pop}
\alias{pop}
\title{German population structure}
\format{
An object of class \code{"data.frame"} with  five columns.
\enumerate{
\item \code{"dist_id"} - An integer representing the county'S unique identifier.
\item \code{"date"} - Date of data publication.
\item \code{"sex"} - Sex of the respective age group.
\item \code{"age_gr"} - The age group.
\item \code{"total"} - Inhabitants of the county in the respective age group and
with the respective sex.
}
}
\source{
\href{https://www.destatis.de/DE/Themen/Laender-Regionen/
Regionales/Gemeindeverzeichnis/Administrativ/Archiv/Standardtabellen/
04_KreiseVorjahr.html}{Statistisches Bundesamt}
}
\usage{
data(pop)
}
\description{
The German population structure on county level (NUTS-3) stratified by age groups
and sex as off 31st of December 2018.
}
\references{
Federal Statistical Office of Germany. Kreisfreie Städte und
Landkreise nach Fläche, Bevölkerung und Bevölkerungsdichte, (2018),
\href{https://www.destatis.de/DE/Themen/Laender-Regionen/Regionales}
{https://www.destatis.de/DE/Themen/Laender-Regionen/Regionales}.
}
\keyword{datasets}
