% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_functions.R
\name{save.exec.params}
\alias{save.exec.params}
\title{Function to save the current list of execution parameters.}
\usage{
save.exec.params(ep)
}
\arguments{
\item{ep}{An execution parameter list as decribed in \code{\link[=set.exec.params]{set.exec.params()}}.}
}
\description{
Function to save the current list of execution parameters.
}
