% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_functions.R
\name{save.input}
\alias{save.input}
\title{Function to save the current list of loaded input data.}
\usage{
save.input(ep, iol)
}
\arguments{
\item{ep}{An execution parameter list as decribed in \code{\link[=set.exec.params]{set.exec.params()}}.}

\item{iol}{A list with loaded input data as described in \code{\link[=load.input]{load.input()}}.}
}
\description{
Function to save the current list of loaded input data.
}
