% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_functions.R
\name{save.optimization.params}
\alias{save.optimization.params}
\title{Function to save the current list of loaded input data.}
\usage{
save.optimization.params(ep, op)
}
\arguments{
\item{ep}{An execution parameter list as decribed in \code{\link[=set.exec.params]{set.exec.params()}}.}

\item{op}{A list with parameters steering the optimization procedure as
described in \code{\link[=set.optimization.params]{set.optimization.params()}}.}
}
\description{
Function to save the current list of loaded input data.
}
