% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_functions.R
\name{save.pspace}
\alias{save.pspace}
\title{Function to save the current psapce list.}
\usage{
save.pspace(ep, pspace)
}
\arguments{
\item{ep}{An execution parameter list as decribed in \code{\link[=set.exec.params]{set.exec.params()}}.}

\item{pspace}{The parameter list pspace.}
}
\description{
Function to save the current psapce list.
}
