% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.population.R
\name{save.spatial.population}
\alias{save.spatial.population}
\title{Function to save the current list of execution parameters.}
\usage{
save.spatial.population(ep, sim.struc)
}
\arguments{
\item{ep}{An execution parameter list as decribed in \code{\link[=set.exec.params]{set.exec.params()}}.}

\item{sim.struc}{List with population data. Use \code{\link[=init.spatial.population]{init.spatial.population()}}
in order to create a valid layout.}
}
\description{
Function to save the current list of execution parameters.
}
