% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_functions.R
\name{save.static.params}
\alias{save.static.params}
\title{Function to save the current list of satic model parameters.}
\usage{
save.static.params(ep, sp)
}
\arguments{
\item{ep}{An execution parameter list as decribed in \code{\link[=set.exec.params]{set.exec.params()}}.}

\item{sp}{A list with static model parameters as described in \code{\link[=set.static.params]{set.static.params()}}.}
}
\description{
Function to save the current list of satic model parameters.
}
