% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infections.R
\docType{data}
\name{seed}
\alias{seed}
\title{Infected cases}
\format{
An object of class \code{"data.frame"} with  ... columns.
\enumerate{
\item \code{"dist_id"} - An integer representing the county's unique identifier.
\item \code{"Name"} - The name of the county.
\item \code{"Area"} - The counties area in km².
\item \code{"Inhabitants"} - The population of the county.
}
}
\source{
\href{https://npgeo-corona-npgeo-de.hub.arcgis.com/datasets/
dd4580c810204019a7b8eb3e0b329dd6_0}{COVID-19 Datenhub}
}
\usage{
data(infections)
}
\description{
Infected cases for seeding infections during model startup.
}
\references{
Robert Koch-Institute, COVID-19 Dashboard, (2020),\cr
\href{https://www.rki.de/DE/Content/InfAZ/N/Neuartiges_Coronavirus/nCoV_node.html}
{https://www.rki.de/DE/Content/InfAZ/N/Neuartiges_Coronavirus/nCoV_node.html}.
}
\keyword{datasets}
