% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_functions.R
\name{set.optimization.params}
\alias{set.optimization.params}
\title{Setup of optimization parameters}
\usage{
set.optimization.params(
  opt.target.icu,
  opt.target.deaths,
  opt.target.region,
  opt.names,
  opt.lb,
  opt.ub,
  opt.pop.size,
  opt.max.iter,
  use.sug.sol,
  ep,
  sp,
  pspc,
  opt.filter = NULL
)
}
\description{
Setup of optimization parameters
}
