% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_functions.R
\name{set.pspace}
\alias{set.pspace}
\title{Setup of parameters in parameter space}
\usage{
set.pspace(param, values, type = "direct", s.dev = NULL)
}
\arguments{
\item{param}{The name of the parameter to be set.}

\item{values}{The values of the parameter to be set.}

\item{type}{The parameter type. Allowed values are \code{direct} or
\code{dist}.\cr
\emph{Defaults to:} \code{direct}}

\item{s.dev}{Deviations of the values in case of parameter type
\code{dist}.\cr
\emph{Defaults to:} \code{NULL}}
}
\value{
The function operates on the global scope and modifies the
parameter list pspace.\
}
\description{
The function adds an element to the parameter space list pspace
}
