% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_functions.R
\name{set.static.params}
\alias{set.static.params}
\title{Setup of static parameters}
\usage{
set.static.params(
  pspace,
  seed.in.inner.loop = FALSE,
  seed.base = NULL,
  country = "Germany",
  restrict = TRUE,
  sim.regions = c("Schleswig-Holstein", "Hamburg", "Niedersachsen", "Bremen"),
  sam_prop.ps = c(1, 1, 1, 1),
  sim_pop = "proportional",
  ini_infected = 10,
  seed_infections = "data",
  seed_date = "2020-03-09",
  seed_before = 7,
  time_n = NULL,
  inf_dur = 3,
  cont_dur = 2,
  ill_dur = 8,
  icu_per_day = c(0, 0, 0, 0, 0, 0, 0, 8),
  less_contagious = 0.7,
  R0_force = 0,
  immune_stop = TRUE,
  import_R0_matrix = FALSE,
  R0change = lapply(seq(1, by = 7, length.out = 20), function(x) {     c(x, x + 6) }),
  R0county = as.list(rep("ALL", 20)),
  R0delay = TRUE,
  R0delay_days = 5,
  R0delay_type = "linear",
  endogenous_lockdown = FALSE,
  lockdown_effect = 0.39,
  lockdown_connect = 0.5,
  lockdown_threshold = 100,
  lockdown_days = 10,
  control_age_sex = "age",
  iter = 4,
  lhc.samples = NULL,
  lhc.reload = FALSE,
  gplots = FALSE,
  cplots = FALSE,
  cplots.states = FALSE,
  cplots.nuts2 = FALSE,
  results = "Reduced",
  sp.states = NULL,
  region = "states"
)
}
\description{
Setup of static parameters
}
