% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{setup.projection}
\alias{setup.projection}
\title{Extraploate 0effects beyond determined values}
\usage{
setup.projection(
  R0effect,
  sp,
  method = "constant-daily",
  base = NULL,
  length = 8,
  length.days = 14
)
}
\arguments{
\item{R0effect}{A data.frame with R0effects per week and region.}

\item{sp}{An object with static CoSMic model parameters.}

\item{method}{Method by which to extrapolate. Supported values are:
\code{"constant-weekly"}: Extrapolates constantly the R0effect
of week \code{base} to the next \code{length} weeks.
\code{"constant-daily"}: Determines the averaged daily R0effect
from the last \code{length.days} and extraploates it constantly
to the next \code{length} weeks.
\emph{Defaults to:} \code{"constant-weekly"}.}

\item{base}{The week based on which to extrapolate. If not given the last
week i.e. dim(R0effect)[1] is used.
\emph{Defaults to:} \code{NULL}.}

\item{length}{Number of week to extrapolate after base.
\emph{Defaults to:} \code{8}.}

\item{length.days}{Number of days to take into account when extraploation
based on daily quantities is done.
\emph{Defaults to:} \code{14}.}
}
\description{
The function extrapolates R0effects based on different methods.
}
