% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/states.R
\docType{data}
\name{states}
\alias{states}
\title{Structure of German federal states}
\format{
An object of class \code{"data.frame"} with  four columns.
\enumerate{
\item \code{"Code"} - An integer representing the states unique identifier.
\item \code{"inhabitants"} - The population of the state.
\item \code{"Shortcut"} - The two letter identifier of the state.
\item \code{"Name"} - The name of the state.
}
}
\source{
\href{https://www.destatis.de/DE/Themen/Laender-Regionen/
Regionales/Gemeindeverzeichnis/Administrativ/Archiv/Standardtabellen/
04_KreiseVorjahr.html}{Statistisches Bundesamt}
}
\usage{
data(states)
}
\description{
The German federal state structure representing the NUTS-1 level for Germany.
}
\references{
Federal Statistical Office of Germany. Kreisfreie Städte und
Landkreise nach Fläche, Bevölkerung und Bevölkerungsdichte, (2018),
\href{https://www.destatis.de/DE/Themen/Laender-Regionen/Regionales}{https://www.destatis.de/DE/Themen/Laender-Regionen/Regionales}.
}
\keyword{datasets}
