% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_pr.R
\docType{data}
\name{trans_pr}
\alias{trans_pr}
\title{Transition probabilities}
\format{
An object of class \code{"data.frame"} with  five columns.
\enumerate{
\item \code{"age_gr"} - Age groups with 5 year stepping from 0 - 90 years.
\item \code{"sex"} - Label for stratification according t sex using labels
"total", "m" and "f"
\item \code{"surv_ill"} - Chance of surviving an infection.
\item \code{"icu_risk"} - Risk for intensive care requirement when infected.
\item \code{"surv_icu"} - Chance of surviving intensive care.
}
}
\usage{
data(trans_pr)
}
\description{
The dataset describes the transition probabilities to intensive care units
along with the chance to survive a Corona virus infction when beeing ill
or being ill and in intensive care stratified by age groups.
}
\keyword{datasets}
