% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_functions.R
\name{convert.Rp.to.Fp}
\alias{convert.Rp.to.Fp}
\title{Convert R-model parameters to Fortran-model input files}
\usage{
convert.Rp.to.Fp(
  filename.sp,
  sp,
  filename.ep,
  ep,
  iol,
  pspace,
  op = NULL,
  outpath = "./"
)
}
\arguments{
\item{sp}{A list with static model parameters as described in \code{\link[=set.static.params]{set.static.params()}}.}

\item{filename}{Path to the output file.}
}
\description{
The function prints the R-model parameter lists and input data to textfiles
which can be used as input for the Fortran model version.
}
