% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoSMic-package.R
\docType{package}
\name{CoSMic-package}
\alias{CoSMic}
\alias{CoSMic-package}
\title{CoSMic: COVID-19 spatial microsimulation for Germany}
\description{
A calibration-microsimulation approach to reduce uncertainty for
    policy decisions on non-pharmacological interventions in the COVID-19
    pandemic.
    The package implements an age-structured spatial microsimulation model that
    extends the Susceptible-Exposed-Infectious-Recovered (SEIR) framework.
    Using an optimization approach based on subnational trends in the number of
    intensive care patients, it is able to calibrate the model to the ongoing
    spread of the epidemic and tries to estimate how the NPIs have affected it.
    Based on these estimates the model can provide national and sub-national
    forecasts for trends in the number of ICU patients and other indicators
    under different scenarios regarding NPIs.
}
\author{
\strong{Maintainer}: Ralf Schneider \email{ralf.schneider@hlrs.de}

Authors:
\itemize{
  \item Christian Dudel \email{dudel@demogr.mpg.de}
  \item Matthias Rosenbaum-Feldbrügge \email{Matthias.Rosenbaum-Feldbruegge@bib.bund.de}
  \item Sebastian Klüsener \email{sebastian.kluesener@bib.bund.de}
}

}
\keyword{package}
