% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosmic.R
\name{cosmic}
\alias{cosmic}
\title{Fit Conditional Ordinal Stereotype Model}
\usage{
cosmic(
  data,
  incidentID,
  officerID,
  y,
  priorSD_lambda = 2,
  priorSD_sDiff = 1,
  iter = 2000,
  chains = 4,
  threads = 2
)
}
\description{
Fit Conditional Ordinal Stereotype Model
}
