% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoSMic.R
\name{CoSMic}
\alias{CoSMic}
\title{Function executing the simulation model.}
\usage{
CoSMic(ep, sp, iol, pspace, sim.struc, op, opt)
}
\arguments{
\item{ep}{Execution parameter list. Use \code{\link[=set.exec.params]{set.exec.params()}} in order to
create a valid layout.}

\item{sp}{List with static model parameters. Use \code{\link[=set.static.params]{set.static.params()}} to
create a valid layout.}

\item{iol}{Input data list. Use \code{\link[=load.input]{load.input()}} to load needed fies and
\code{\link[=init.connectivity]{init.connectivity()}} in order to create a valid date layout.}

\item{pspace}{List holding the parameter space with potentially variable
model parameters. Use the setter function \code{\link[=set.pspace]{set.pspace()}} to
add parameters.}

\item{sim.struc}{List with population data. Use \code{\link[=init.spatial.population]{init.spatial.population()}}
in order to create a valid layout.}

\item{op}{List with steering parameters for the optimization process. \cr
Use \code{\link[=set.optimization.params]{set.optimization.params()}} in order to create a valid layout
and\cr \code{\link[=init.reference.data]{init.reference.data()}} in order to init the optimization
targets based on observed data.}

\item{opt}{Numeric vector with model parameters subject to optimization.}
}
\value{
Depends upon the selected execution procedure given by
ep$exec.procedure.
\enumerate{
\item{In case \code{ep$exec.procedure="Optimization"} a scalar
target value is returned.}
\item{In case \code{ep$exec.procedure="Basic-Param"} a list
with transient result data is returned.}}
}
\description{
Function executing the simulation model.
}
\section{ToDo}{

\itemize{
\item{Capture Error Messages in foreach and model loop}
\item{Fix county plots}
\item{Implement statisitcs output against opt.targets}
\item{Implement normed standard deviation as target value in
Global daths & icu_cases & local deaths}
\item{Implement Error message in case R0county contains county id which
is not selected for simulation.}
}
}

