% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_functions.R
\name{load.input}
\alias{load.input}
\title{Loading input data}
\usage{
load.input(
  data.dir = "./",
  trans.pr = NULL,
  pop.data = NULL,
  inf.cases = NULL,
  dead.cases = NULL,
  connect.total = NULL,
  connect.work = NULL,
  sts = NULL,
  cnts = NULL,
  R0.matrix.inp = NULL,
  dead.cases.by.state = NULL,
  dead.cases.by.country = NULL,
  icu.cases.by.county = NULL,
  icu.cases.by.state = NULL,
  icu.cases.by.country = NULL,
  lhc.data = NULL
)
}
\description{
Loading input data
}
