% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_functions.R
\name{set.static.params}
\alias{set.static.params}
\title{Setup of static parameters}
\usage{
set.static.params(
  pspace,
  seed.in.inner.loop,
  seed.base,
  country,
  restrict,
  sim.regions,
  sam_prop.ps,
  sim_pop,
  ini_infected,
  seed_infections,
  seed_date,
  seed_before,
  time_n = NULL,
  inf_dur,
  cont_dur,
  ill_dur,
  icu_per_day,
  less_contagious,
  R0_force,
  immune_stop,
  import_R0_matrix = FALSE,
  R0change,
  R0county,
  R0delay,
  R0delay_days,
  R0delay_type,
  endogenous_lockdown,
  lockdown_effect,
  lockdown_connect,
  lockdown_threshold,
  lockdown_days,
  control_age_sex,
  iter = NULL,
  lhc.samples = NULL,
  lhc.reload = FALSE,
  gplots = FALSE,
  cplots = FALSE,
  cplots.states = FALSE,
  cplots.nuts2 = FALSE,
  results = "Reduced",
  sp.states = NULL
)
}
\description{
Setup of static parameters
}
